refactor(layout): split repeated chrome detection
Move cross-page repeated chrome marking and shared layout geometry helpers out of the main layout builder so the page layout pipeline can be reviewed separately from repeated-header cleanup.
